rag-example/requirements.txt
llama-index==0.11.4
PyPDF2
sentence-transformers==2.6.1
pinecone-client
mistralai
llama-index-embeddings-mistralai
llama-index-llms-mistralai
pycryptodome
llama-index-vector-stores-pinecone==0.2.1
llama-index-readers-file==0.2.0